+++ src/ALGraphV2.cpp
@@ -9,17 +9,17 @@
 
 void ALGraphV2::add_edge(int from, std::vector<int>& to, std::vector<int>& w) {
     for(int i : to)
-        edges[from].push_back(i);
+        edges[from].push_back((uint64_t)i);
     for(int i : w)
-        weights[from].push_back(i);
+        weights[from].push_back((uint64_t)i);
 }
 
 void ALGraphV2::finished() {
 }
 
 void ALGraphV2::add_edge(int from, int to, int weight) {
-    edges[from].push_back(to);
-    weights[from].push_back(weight);
+    edges[from].push_back((uint64_t)to);
+    weights[from].push_back((uint64_t)weight);
 }
 
 void ALGraphV2::sortByEdgesByNodeId() {
